requirements.txt
mcp[cli]
pypdf
python-pptx
ollama
